Refactor codebase and restructure profile files
