+++ src/routes/tasks.route.js
@@ -1,0 +1,9 @@
+
+
+export default (router) => {
+
+	/** 
+	 * TODO: - Add all other routes for tasks
+	 */
+	
+}
